+++ opennmt/utils/checkpoint.py
@@ -54,8 +54,9 @@ def _update_vocabulary_variables(variables, current_vocab_path, new_vocab_path, 
       tf.logging.debug("Updating variable %s" % name)
       variables[name] = _update_vocabulary_variable(tensor, current_size, mapping)
 
-def _save_new_variables(variables, output_dir, base_checkpoint_path, session_config=None):
+def _create_checkpoint_from_variables(variables, output_dir, latest_step=None, session_config=None):
   if "global_step" in variables:
+    latest_step = variables["global_step"]
     del variables["global_step"]
   tf_vars = []
   for name, value in six.iteritems(variables):
@@ -72,7 +73,6 @@ def _save_new_variables(variables, output_dir, base_checkpoint_path, session_con
   placeholders = [tf.placeholder(v.dtype, shape=v.shape) for v in tf_vars]
   assign_ops = [tf.assign(v, p) for (v, p) in zip(tf_vars, placeholders)]
 
-  latest_step = int(base_checkpoint_path.split("-")[-1])
   out_base_file = os.path.join(output_dir, "model.ckpt")
   global_step = tf.get_variable(
       "global_step",
@@ -88,6 +88,52 @@ def _save_new_variables(variables, output_dir, base_checkpoint_path, session_con
     saver.save(sess, out_base_file, global_step=global_step)
 
   return output_dir
+
+def get_checkpoint_variables(checkpoint_path):
+  """Returns variables included in a checkpoint.
+
+  Args:
+    checkpoint_path: Path to the checkpoint.
+
+  Returns:
+    A dictionary mapping variables name to value.
+  """
+  reader = tf.train.load_checkpoint(checkpoint_path)
+  return {
+      name:reader.get_tensor(name)
+      for name in six.iterkeys(reader.get_variable_to_shape_map())}
+
+def convert_checkpoint(checkpoint_path,
+                       output_dir,
+                       source_dtype,
+                       target_type,
+                       session_config=None):
+  """Converts checkpoint variables from one dtype to another.
+
+  Args:
+    checkpoint_path: The path to the checkpoint to convert.
+    output_dir: The directory that will contain the converted checkpoint.
+    source_dtype: The data type to convert from.
+    target_dtype: The data type to convert to.
+    session_config: Optional configuration to use when creating the session.
+
+  Returns:
+    The path to the directory containing the converted checkpoint.
+
+  Raises:
+    ValueError: if :obj:`output_dir` points to the same directory as
+      :obj:`checkpoint_path`.
+  """
+  if os.path.dirname(checkpoint_path) == output_dir:
+    raise ValueError("Checkpoint and output directory must be different")
+  variables = get_checkpoint_variables(checkpoint_path)
+  for name, value in six.iteritems(variables):
+    if not name.startswith("optim") and tf.as_dtype(value.dtype) == source_dtype:
+      variables[name] = value.astype(target_type.as_numpy_dtype())
+  return _create_checkpoint_from_variables(
+      variables,
+      output_dir,
+      session_config=session_config)
 
 def update_vocab(model_dir,
                  output_dir,
@@ -130,17 +176,14 @@ def update_vocab(model_dir,
     return model_dir
   checkpoint_path = tf.train.latest_checkpoint(model_dir)
   tf.logging.info("Updating vocabulary related variables in checkpoint %s" % checkpoint_path)
-  reader = tf.train.load_checkpoint(checkpoint_path)
-  variable_map = reader.get_variable_to_shape_map()
-  variable_value = {name:reader.get_tensor(name) for name, _ in six.iteritems(variable_map)}
+  variable_value = get_checkpoint_variables(checkpoint_path)
   if new_src_vocab is not None:
     _update_vocabulary_variables(variable_value, current_src_vocab, new_src_vocab, "encoder", mode)
   if new_tgt_vocab is not None:
     _update_vocabulary_variables(variable_value, current_tgt_vocab, new_tgt_vocab, "decoder", mode)
-  return _save_new_variables(
+  return _create_checkpoint_from_variables(
       variable_value,
       output_dir,
-      checkpoint_path,
       session_config=session_config)
 
 
@@ -199,8 +242,9 @@ def average_checkpoints(model_dir, output_dir, max_count=8, session_config=None)
     for name in avg_values:
       avg_values[name] += reader.get_tensor(name) / num_checkpoints
 
-  return _save_new_variables(
+  latest_step = int(checkpoints_path[-1].split("-")[-1])
+  return _create_checkpoint_from_variables(
       avg_values,
       output_dir,
-      checkpoints_path[-1],
+      latest_step=latest_step,
       session_config=session_config)